backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
google-genai==1.44.0
google-auth==2.38.0
python-dotenv==1.0.0
websockets==15.0.1
slowapi==0.1.9
redis==7.1.0
